Re-organize components stories
