+++ stories/components/alert.stories.js
@@ -1,6 +1,6 @@
 import { radios } from '@storybook/addon-knobs'
 import { withKnobs } from '@storybook/addon-knobs'
-import { insertTemplateOnContainer } from '../.storybook/helpers'
+import { insertTemplateOnContainer } from '../../.storybook/helpers'
 
 export default {
   title: 'Components/Alert',
