+++ stories/components/button.stories.js
@@ -1,6 +1,6 @@
 import { text, radios, boolean } from '@storybook/addon-knobs'
 import { withKnobs } from '@storybook/addon-knobs'
-import { insertTemplateOnContainer } from '../.storybook/helpers'
+import { insertTemplateOnContainer } from '../../.storybook/helpers'
 
 export default {
   title: 'Components/Buttons',
